# docs/architecture-pack/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Farmmap API
  version: 1.0.0
  description: >
    Farmmap public and authenticated API covering all Must-Have features (F1–F16).
    All endpoints require a Supabase JWT Bearer token where marked. directory_id
    is resolved server-side from the Host header; it is never a client parameter
    on public endpoints.
  contact:
    name: Farmmap Engineering
    url: https://farmmap.co.uk
servers:
  - url: https://farmmap.co.uk/api
    description: Production
  - url: https://staging.farmmap.co.uk/api
    description: Staging
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT token from auth.getSession()
  parameters:
...
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: not_found
        message:
          type: string
          example: Listing not found
        details:
          type: object
          additionalProperties: true
    ListingType:
      type: string
      enum: [farm_shop, honesty_box, farm_gate_stall, roadside_stand]
    ListingStatus:
      type: string
      enum: [pending, active, claimed, suspended]
    SubscriptionTier:
      type: string
      enum: [free, bronze, silver, gold]
...
    GeoPoint:
      type: object
      required: [lon, lat]
      properties:
        lon:
          type: number
          format: double
          minimum: -180
          maximum: 180
        lat:
          type: number
          format: double
          minimum: -90
          maximum: 90
...
    ListingSummary:
      type: object
      required: [id, slug, name, listing_type, status, tier, town, county, country, location]
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        listing_type:
          $ref: '#/components/schemas/ListingType'
        status:
          $ref: '#/components/schemas/ListingStatus'
        tier:
          $ref: '#/components/schemas/SubscriptionTier'
        town:
          type: string
        county:
          type: string
        country:
          type: string
        location:
          $ref: '#/components/schemas/GeoPoint'
        review_rating_avg:
          type: number
          format: float
          nullable: true
        review_count:
          type: integer
        hero_photo_url:
          type: string
          format: uri
          nullable: true
...
    OrderStatus:
      type: string
      enum: [pending, accepted, preparing, dispatched, delivered, cancelled, refunded]
    OrderItem:
      type: object
      required: [product_id, product_name, quantity, unit_price_pence, line_total_pence]
      properties:
        product_id:
          type: string
          format: uuid
        product_name:
          type: string
        quantity:
          type: integer
          minimum: 1
        unit_price_pence:
          type: integer
        line_total_pence:
          type: integer
    Order:
      type: object
      required: [id, status, subtotal_pence, total_pence, items, created_at]
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/OrderStatus'
        subtotal_pence:
          type: integer
        commission_pence:
          type: integer
        delivery_fee_pence:
          type: integer
        total_pence:
          type: integer
        currency:
          type: string
          enum: [GBP, EUR]
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        delivery_slot:
          type: object
          nullable: true
          properties:
            date:
              type: string
              format: date
            time_from:
              type: string
            time_to:
              type: string
        notes:
          type: string
          nullable: true
        auto_cancel_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
...
    Pagination:
      type: object
      required: [page, limit, total, total_pages]
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnprocessableEntity:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /listings:
    get:
      operationId: listListings
      summary: Search and filter listings
      description: >
        Returns listings matching the given filters. Supports geospatial radius
        search, bounding box, listing type, product category, and open/closed filter.
        directory_id is resolved server-side from the Host header.
      tags: [Listings]
      parameters:
        - name: lat
          in: query
          schema:
            type: number
            format: double
          description: Latitude for radius search
        - name: lon
          in: query
          schema:
            type: number
            format: double
          description: Longitude for radius search
        - name: radius_metres
          in: query
          schema:
            type: integer
            minimum: 100
            maximum: 100000
            default: 25000
        - name: bbox
          in: query
          schema:
            type: string
            pattern: '^-?\d+\.?\d*,-?\d+\.?\d*,-?\d+\.?\d*,-?\d+\.?\d*$'
          description: Bounding box as minLon,minLat,maxLon,maxLat for map viewport queries
        - name: listing_type
          in: query
          schema:
            $ref: '#/components/schemas/ListingType'
        - name: product_category
          in: query
          schema:
            type: string
        - name: open_now
          in: query
          schema:
            type: boolean
          description: Filter to listings currently open based on declared opening hours
        - name: tier
          in: query
          schema:
            $ref: '#/components/schemas/SubscriptionTier'
        - name: county
          in: query
          schema:
            type: string
        - name: country
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: q
          in: query
          schema:
            type: string
            maxLength: 200
          description: Free text search against listing name and description
      responses:
        '200':
          description: Listings matching the query
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ListingSummary'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /listings/geojson:
    get:
      operationId: listingsGeoJSON
      summary: GeoJSON endpoint for MapLibre map pins
      description: >
        Returns all active listings as a GeoJSON FeatureCollection with minimal
        properties per feature. Used by MapLibre GL JS to render the map.
        Optionally filtered by bounding box for viewport-based loading.
      tags: [Listings]
      parameters:
        - name: bbox
          in: query
          schema:
            type: string
        - name: listing_type
          in: query
          schema:
            $ref: '#/components/schemas/ListingType'
        - name: product_category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: GeoJSON FeatureCollection of active listing pins
          content:
            application/geo+json:
              schema:
                type: object
                required: [type, features]
                properties:
                  type:
                    type: string
                    enum: [FeatureCollection]
                  features:
                    type: array
                    items:
                      type: object
                      properties:
                        type:
                          type: string
                          enum: [Feature]
                        geometry:
                          type: object
                          properties:
                            type:
                              type: string
                              enum: [Point]
                            coordinates:
                              type: array
                              items:
                                type: number
                              minItems: 2
                              maxItems: 2
                        properties:
                          type: object
                          properties:
                            id:
                              type: string
                              format: uuid
                            slug:
                              type: string
                            name:
                              type: string
                            listing_type:
                              $ref: '#/components/schemas/ListingType'
                            tier:
                              $ref: '#/components/schemas/SubscriptionTier'
                            rating_avg:
                              type: number
                              nullable: true
                            temporarily_closed:
                              type: boolean
...
  /orders/{orderId}/status:
    patch:
      operationId: updateOrderStatus
      summary: Update order status
      description: >
        Owner/manager transitions order through the status workflow:
        pending → accepted → preparing → dispatched → delivered.
        Also handles: cancellation (owner), refund initiation (owner).
        Consumer is notified on each transition. Completing the order (delivered)
        increments listing_subscriptions.completed_order_count.
      tags: [Orders]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  $ref: '#/components/schemas/OrderStatus'
                cancellation_reason:
                  type: string
                  maxLength: 500
      responses:
        '200':
          description: Order status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid status transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================
  # MANAGERS (F16 — multi-manager from v1)
  # ============================================================
...